refactor(extract/types): move the apple version comparator into apple.go
